Add explicit cast to Uint64 in IMG_anim_encoder.c

+++ src/IMG_anim_encoder.c
@@ -196,10 +196,10 @@ bool IMG_CloseAnimationEncoder(IMG_AnimationEncoder *encoder)
 
 Uint64 IMG_GetResolvedDuration(IMG_AnimationEncoder* encoder, Uint64 duration, int factor)
 {
-    return SDL_round((encoder->accumulated_pts + duration) * factor * encoder->timebase_numerator / encoder->timebase_denominator) - SDL_round((encoder->accumulated_pts * factor) * encoder->timebase_numerator / encoder->timebase_denominator);
+    return (Uint64)SDL_round((encoder->accumulated_pts + duration) * factor * encoder->timebase_numerator / encoder->timebase_denominator) - SDL_round((encoder->accumulated_pts * factor) * encoder->timebase_numerator / encoder->timebase_denominator);
 }
 
 Uint64 IMG_GetCurrentTimestamp(IMG_AnimationEncoder* encoder, int factor)
 {
-    return SDL_round((encoder->accumulated_pts + encoder->last_delay) * factor * encoder->timebase_numerator / encoder->timebase_denominator);
+    return (Uint64)SDL_round((encoder->accumulated_pts + encoder->last_delay) * factor * encoder->timebase_numerator / encoder->timebase_denominator);
 }
